Changed the project structure and added folders to all components and move styling files to styles etc
