+++ frontend/package.json
@@ -16,7 +16,7 @@
     "@mapbox/mapbox-gl-geocoder": "5.0.2",
     "@mui/icons-material": "5.15.8",
     "@mui/joy": "5.0.0-beta.29",
-    "@mui/material": "^5.15.10",
+    "@mui/material": "5.15.10",
     "@mui/x-date-pickers": "6.19.4",
     "mapbox-gl": "1.13.2",
     "mapbox-gl-geocoder": "2.0.1",
