+++ frontend/src/App.tsx
@@ -1,16 +1,16 @@
 // import './App.css';
 import { Route, Routes } from 'react-router-dom';
 import Login from '@/pages/Login';
-import HandleLogin from '@/pages/HandleLogin';
-import NotFound from '@/pages/NotFound';
-import Success from '@/pages/Token';
+import HandleLogin from '@/utils/Login/HandleLogin';
+import Success from '@/utils/Authorization/Token';
 import Error from '@/pages/Error';
-import { Activities } from './pages/Activities';
+import { Home } from './pages/Home';
+import NotFound from '@/pages/NotFound';
 import { QueryClient, QueryClientProvider } from 'react-query';
 import Activity from './components/RegisterActivity/Activity';
-import ScheduleForm from './components/CreateScheduleComponent';
-import ExportFileUI from './pages/ExportFileUI';
-import TheMenu from './pages/TheMenu';
+import ScheduleForm from '@/components/CreateSchedule/CreateScheduleComponent';
+import ExportFileUI from './components/ExportSchedule/ExportFileUI';
+import TheMenu from './components/HeaderMenu/TheMenu';
 
 // Create a client
 const queryClient = new QueryClient();
@@ -25,7 +25,7 @@ function App() {
         <Route path='/token' element={<Success />} />
         <Route path='/error' element={<Error />} />
         <Route path='*' element={<NotFound />} />
-        <Route path='/activities' element={<Activities />} />
+        <Route path='/home' element={<Home />} />
         <Route path='/activity' element={<Activity />} />
         <Route path='/postschedule' element={<ScheduleForm />} />
         <Route path='/ui' element={<ExportFileUI />} />
