--- frontend/src/components/Common/colors.tsx
@@ -1,7 +1,0 @@
-import { colors } from '@kokitotsos/react-components';
-
-export const primary = colors.primary;
-export const primaryText = '#e5e5e5';
-export const primaryLink = colors.link;
-
-export const secondaryLink = colors.text1;
