+++ frontend/src/components/MapBox/MapBox.tsx
@@ -2,7 +2,7 @@ import { useEffect, useState, useRef } from 'react';
 import mapboxgl from 'mapbox-gl';
 import MapboxGeocoder from '@mapbox/mapbox-gl-geocoder';
 import '@mapbox/mapbox-gl-geocoder/dist/mapbox-gl-geocoder.css';
-import '../MapBox/MapBox.css';
+import '@/styles/MapBox/MapBox.css';
 import { StyledButton, StyledLine } from '../RegisterActivity/StyledActivity';
 
 const MapBox = ({
