--- frontend/src/pages/Root.tsx
@@ -1,5 +1,0 @@
-const Root = () => {
-  return <div>Root</div>;
-};
-
-export default Root;
