+++ frontend/src/styles/Common/colors.tsx
@@ -1,0 +1,4 @@
+export enum Colors {
+  primary = '#f4f4f4',
+  primaryText = '#e5e5e5',
+}
