+++ frontend/src/styles/ExpandInfo/StyledExpandInfo.tsx
